Converted to Streaming Architecture
In the pursuit of the ever elusive sslConfig.config.loos.disableSNI=false being printed to the the log I have done a major overhaul which has made the program faster. No real reason other than this little println somehwere must be hunted down for the good of humanity. These are the important things in life.